Refactored code to src/ and src/commands/
